librus/pyproject.toml
[project]
name = "librus-messages"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "requests",
]
